Input Re-write
The previous iteration of the input was not working, so I re-wrote to check status every iteration of the loop. However, this is not fast enough. I have to hold the button down for a second in order to trigger the event. Therefore, I think I need an interupt to place inputs in the queue. Then I will process those inputs once per frame.